news-processor/requirements.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
feedparser==6.0.11
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
